add es and mysql storage

+++ src/main/java/cn/edu/cqupt/clustering/view/SpectrumTable.java
@@ -78,8 +78,8 @@ public class SpectrumTable {
         spectrumTableView.getColumns().add(seqCol);
 
         // charge
-        TableColumn<Spectrum, Float> chaCol = new TableColumn<>("Charge");
-        chaCol.setCellValueFactory(new PropertyValueFactory<Spectrum, Float>("Charge"));
+        TableColumn<Spectrum, Integer> chaCol = new TableColumn<>("Charge");
+        chaCol.setCellValueFactory(new PropertyValueFactory<Spectrum, Integer>("Charge"));
         chaCol.setStyle("-fx-font-family: 'Arial'; -fx-font-size: 14; -fx-alignment: CENTER");
         spectrumTableView.getColumns().add(chaCol);
